+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - [Redo stack is no longer lost when interacting with text literals][11908].
 - [Fixed bug when clicking header in Table Editor Widget didn't start editing
   it][12064]
+- [Fixed bugs occurring after renaming project from within graph editor][12106].
 
 [11889]: https://github.com/enso-org/enso/pull/11889
 [11836]: https://github.com/enso-org/enso/pull/11836
@@ -18,6 +19,7 @@
 [11902]: https://github.com/enso-org/enso/pull/11902
 [11908]: https://github.com/enso-org/enso/pull/11908
 [12064]: https://github.com/enso-org/enso/pull/12064
+[12106]: https://github.com/enso-org/enso/pull/12106
 
 #### Enso Standard Library